+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/impl/controlflow/If.java
@@ -179,8 +179,8 @@ public class If extends DifferentialFunction implements CustomOp {
     }
 
     @Override
-    public int[] iArgs() {
-        return new int[0];
+    public long[] iArgs() {
+        return new long[0];
     }
 
     @Override
@@ -204,7 +204,7 @@ public class If extends DifferentialFunction implements CustomOp {
     }
 
     @Override
-    public Integer getIArgument(int index) {
+    public Long getIArgument(int index) {
         return null;
     }
