+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/impl/controlflow/While.java
@@ -486,8 +486,8 @@ public class While extends DifferentialFunction implements CustomOp {
     }
 
     @Override
-    public int[] iArgs() {
-        return new int[0];
+    public long[] iArgs() {
+        return new long[0];
     }
 
     @Override
@@ -511,7 +511,7 @@ public class While extends DifferentialFunction implements CustomOp {
     }
 
     @Override
-    public Integer getIArgument(int index) {
+    public Long getIArgument(int index) {
         return null;
     }
